graphql_n_broker/item-service/app/requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
kafka-python==2.0.2
pydantic==1.7.3
python-dotenv==0.15.0
PyYAML==5.4.1
starlette==0.13.6
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
